3011 Refactoring

+++ src/main/java/com/parking/model/ParkedVehicle.java
@@ -1,5 +1,6 @@
 package com.parking.model;
 
+import com.parking.dto.Status;
 import java.time.Instant;
 import javax.persistence.*;
 
@@ -8,7 +9,7 @@ public class ParkedVehicle {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String licensePlate;
 
@@ -31,11 +32,11 @@ public class ParkedVehicle {
     @ManyToOne
     private RouteCycle routeCycle;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
